moved mgcg into doubly_bounded_poisson
